pyproject.toml
[tool.poetry]
name = "zaim-api-sample"
version = "0.1.0"
description = ""
authors = ["Tatsu015 <tatsumi.soichiro@gmail.com>"]
readme = "README.md"
packages = [{include = "zaim_api_sample"}]

[tool.poetry.dependencies]
python = "^3.10"
pyzaim = "^1.0.11"
python-dotenv = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
